+++ test/test_ul4.py
@@ -3665,7 +3665,7 @@ def test_nestedscopes(r):
 	assert "1" == r(source, keepws=False)
 
 
-	# Subtemplates see themselves (i.e. the ``TemplateClosure`` object created for them), but no variables defined later
+	# Subtemplates don't see themselves (i.e. the ``TemplateClosure`` object created for them)
 	source = """
 	<?def x?>
 		<?print type(x)?>;<?print type(y)?>
@@ -3673,7 +3673,7 @@ def test_nestedscopes(r):
 	<?code y = 42?>
 	<?code x.render()?>
 	"""
-	assert "template;undefined" == r(source, keepws=False)
+	assert "undefined;undefined" == r(source, keepws=False)
 
 	# This shows the difference between local variables and variables from the parent.
 	# ``x`` is passed to the subtemplate, so it will always be the current value instead of the one when it is defined