+++ NEWS.rst
@@ -1,3 +1,11 @@
+Changes in 5.13 (released ??/??/2014)
+-------------------------------------
+
+*	The option :option:`setproctitle` for sisyphus jobs has been renamed to
+	:option:`proctitle`. The new method :meth:`setproctitle` sets the process
+	title and can be used to customize setting the process title.
+
+
 Changes in 5.12 (released 11/07/2014)
 -------------------------------------